+++ src/com/cyber/kinoost/db/repositories/MusicRepository.java
@@ -19,10 +19,14 @@ public class MusicRepository extends FilmMusicRepository {
 	}
 	
 	public void createMusic(Music music) {
+		if(music == null) return;
+		
 		musicDao.create(music);
 	}
 	
 	public void createMusicList(final List<Music> music) {
+		if(music == null) return;
+		
         musicDao.callBatchTasks(new Callable<Void>() {
 	        @Override
 	        public Void call() throws Exception {
@@ -35,10 +39,14 @@ public class MusicRepository extends FilmMusicRepository {
 	}
 	
 	public void editMusic(Music music) {
+		if(music == null) return;
+		
 		musicDao.createOrUpdate(music);
 	}
 	
 	public void editMusicList(final List<Music> music) {
+		if(music == null) return;
+		
         musicDao.callBatchTasks(new Callable<Void>() {
 	        @Override
 	        public Void call() throws Exception {
@@ -51,10 +59,14 @@ public class MusicRepository extends FilmMusicRepository {
 	}
 	
 	public void deleteMusic(Music music) {
+		if(music == null) return;
+		
 		musicDao.delete(music);
 	}
 	
 	public void deleteMusicList(List<Music> music) {
+		if(music == null) return;
+		
 		musicDao.delete(music);
 	}
 }
